refactor(marazion): unify hydrology indexing on HydrologyComplex
Drop `PreparedHydroComplex` and rename `WatershedDepressionComplex` to
`HydrologyComplex`, so one type owns indexed nodes plus carve → rim → apron.
Terrain now applies a single `ComposedElevationOp::Hydrology` instead of staged
carve/rim/apron ops, which keeps iteration on correction math in one place.

## Agent Dialogue

- **Developer:** Unify by dropping `PreparedHydroComplex`, renaming the complex to `HydrologyComplex`, and implementing `ElevationModulation` on that type directly (stop staging via `ComposedElevationOp` carve/rim/apron variants).
- **Agent:** Moved indexing + sample eval onto `HydrologyComplex`; `ComposedElevationOp` is now `Jersey | Hydrology`; Terrain pulls one complex op; stage cells remain as bones. `cargo test -p marazion-watersheds` and `cargo check -p durham-terrain-models` green.
