+++ maybraid/durham/models/src/terrain/jersey_modulation.rs
@@ -2,26 +2,21 @@
 
 use crate::terrain::sdf::{ElevationModulation, TerrainSdf};
 use jersey_terrain_stamps::JerseyModulation;
-use marazion_watersheds::{CorrectionStage, PreparedHydroComplex};
+use marazion_watersheds::HydrologyComplex;
 
 /// One elevation op in the final terrain stack.
-///
-/// Watershed correction is staged: carve → rim → apron (not a closed leaf op).
 #[derive(Debug, Clone)]
 pub enum ComposedElevationOp {
 	Jersey(JerseyModulation),
-	WatershedCarve(PreparedHydroComplex),
-	WatershedRim(PreparedHydroComplex),
-	WatershedApron(PreparedHydroComplex),
+	/// Indexed hydrology complex (internally carve → rim → apron).
+	Hydrology(HydrologyComplex),
 }
 
 impl ComposedElevationOp {
 	pub fn modify_elevation_xz(&self, elevation: f32, x: f32, z: f32) -> f32 {
 		match self {
 			Self::Jersey(m) => JerseyModulation::modify_elevation(m, elevation, x, z),
-			Self::WatershedCarve(h) => h.apply_stage(CorrectionStage::Carve, elevation, x, z),
-			Self::WatershedRim(h) => h.apply_stage(CorrectionStage::Rim, elevation, x, z),
-			Self::WatershedApron(h) => h.apply_stage(CorrectionStage::Apron, elevation, x, z),
+			Self::Hydrology(h) => HydrologyComplex::modify_elevation(h, elevation, x, z),
 		}
 	}
 }
@@ -52,7 +47,7 @@ impl ElevationModulation for ComposedElevationOp {
 	}
 }
 
-impl ElevationModulation for PreparedHydroComplex {
+impl ElevationModulation for HydrologyComplex {
 	fn modify_elevation(
 		&self,
 		_terrain: &TerrainSdf,
@@ -61,7 +56,6 @@ impl ElevationModulation for PreparedHydroComplex {
 		z: f32,
 		_index: usize,
 	) -> f32 {
-		// Full stack only for ad-hoc sampling; terrain uses staged ops.
-		PreparedHydroComplex::modify_elevation(self, elevation, x, z)
+		HydrologyComplex::modify_elevation(self, elevation, x, z)
 	}
 }
